requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
asttokens==2.4.1
    # via stack-data
backcall==0.2.0
    # via ipython
black==24.2.0
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
decorator==5.1.1
    # via ipython
executing==2.0.1
    # via stack-data
flake8==7.0.0
    # via -r requirements.in
grpcio==1.60.1
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.60.1
    # via -r requirements.in
idna==3.6
    # via requests
importlib-metadata==7.0.1
    # via build
ipython==8.12.3
    # via -r requirements.in
jedi==0.19.1
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via
    #   black
    #   build
parso==0.8.3
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==7.4.0
    # via -r requirements.in
platformdirs==4.2.0
    # via black
prompt-toolkit==3.0.43
    # via ipython
protobuf==4.25.3
    # via grpcio-tools
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.11.1
    # via flake8
pydantic==2.6.1
    # via -r requirements.in
pydantic-core==2.16.2
    # via pydantic
pyflakes==3.2.0
    # via flake8
pygments==2.17.2
    # via ipython
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
python-decouple==3.8
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
six==1.16.0
    # via asttokens
stack-data==0.6.3
    # via ipython
tomli==2.0.1
    # via
    #   black
    #   build
    #   pip-tools
    #   pyproject-hooks
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.9.0
    # via
    #   annotated-types
    #   black
    #   ipython
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.42.0
    # via pip-tools
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
